+++ extensions/brave-search-with-results/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Brave Search with Results Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Image search
+- Use common keyboard shortcuts
+
 ## [Initial Version] - 2024-05-04
 
 - Web results